+++ CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/), and this project adheres to [Semantic Versioning](https://semver.org/).
+## [Unreleased]
+
+### Added
++ New-Exception ([#7](https://github.com/PoshAJ/PoshToolbox/pull/7))
+    > Creates an instance of ErrorRecord.
++ Split-File ([#11](https://github.com/PoshAJ/PoshToolbox/pull/11))
+    > Splits large files into several smaller files.
++ Join-File ([#11](https://github.com/PoshAJ/PoshToolbox/pull/11))
+    > Combines split files into their larger source file.
+
+### Fixed
++ Typos and Formatting ([#12](https://github.com/PoshAJ/PoshToolbox/pull/12))
+
+### Changed
++ Resolve-PoshPath ([#9](https://github.com/PoshAJ/PoshToolbox/pull/9))
+    > Output changed to result object, similar to the Resolve-Path cmdlet.
++ Get-FolderProperties ([#9](https://github.com/PoshAJ/PoshToolbox/pull/9))
+    > Refactor for Resolve-PoshPath changes.
++ Start-PoshLog ([#13](https://github.com/PoshAJ/PoshToolbox/pull/12))
+    > Refactor for Resolve-PoshPath changes.
++ Write-PoshLog ([#13](https://github.com/PoshAJ/PoshToolbox/pull/12))
+    > Refactor for Resolve-PoshPath changes.
++ Stop-PoshLog ([#13](https://github.com/PoshAJ/PoshToolbox/pull/12))
+    > Refactor for Resolve-PoshPath changes.
 
 ## [2.0.0] - 2023-04-02
 Project name change, previous versions known as `ScriptFramework`.